Skip testing Qwen3 ASR for ubuntu-latest

+++ .github/scripts/test-dart.sh
@@ -6,9 +6,11 @@ cd dart-api-examples
 
 pushd non-streaming-asr
 
-echo '----------Qwen3 ASR----------'
-./run-qwen3-asr.sh
-rm -rf sherpa-onnx-*
+if [[ $SKIP_QWEN3 != "true" ]]; then
+  echo '----------Qwen3 ASR----------'
+  ./run-qwen3-asr.sh
+  rm -rf sherpa-onnx-*
+fi
 
 echo '----------Moonshine v2----------'
 ./run-moonshine-v2.sh
