+++ .github/scripts/test-dot-net.sh
@@ -4,19 +4,23 @@ set -ex
 
 cd dotnet-examples/
 
-cd ./vad-non-streaming-qwen3-asr
-./run-ten-vad.sh
-rm -fv *.onnx
+if [[ $SKIP_QWEN3 != "true" ]]; then
+  cd ./vad-non-streaming-qwen3-asr
+  ./run-ten-vad.sh
+  rm -fv *.onnx
 
-./run.sh
-rm -fv *.onnx
+  ./run.sh
+  rm -fv *.onnx
 
-cd ../non-streaming-qwen3-asr-decode-files
-./run.sh
-ls -lh
-rm -rf sherpa-onnx-qwen3-*
+  cd ../non-streaming-qwen3-asr-decode-files
+  ./run.sh
+  ls -lh
+  rm -rf sherpa-onnx-qwen3-*
 
-cd ../source-separation-spleeter
+  cd ..
+fi
+
+cd ./source-separation-spleeter
 ./run.sh
 ls -lh
 rm -rf sherpa-onnx-spleeter-*
