+++ .github/scripts/test-python.sh
@@ -8,22 +8,24 @@ log() {
   echo -e "$(date '+%Y-%m-%d %H:%M:%S') (${fname}:${BASH_LINENO[0]}:${FUNCNAME[1]}) $*"
 }
 
-log "test Qwen3 ASR"
+if [[ $SKIP_QWEN3 != "true" ]]; then
+  log "test Qwen3 ASR"
 
-wget -q https://github.com/k2-fsa/sherpa-onnx/releases/download/asr-models/sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25.tar.bz2
-tar xvf sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25.tar.bz2
-rm sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25.tar.bz2
+  wget -q https://github.com/k2-fsa/sherpa-onnx/releases/download/asr-models/sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25.tar.bz2
+  tar xvf sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25.tar.bz2
+  rm sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25.tar.bz2
 
-python3 ./python-api-examples/offline-qwen3-asr-decode-files.py \
-  --conv-frontend=./sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25/conv_frontend.onnx \
-  --encoder=./sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25/encoder.int8.onnx \
-  --decoder=./sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25/decoder.int8.onnx \
-  --tokenizer=./sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25/tokenizer \
-  --max-new-tokens=128 \
-  --num-threads=2 \
-  ./sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25/test_wavs/raokouling.wav
+  python3 ./python-api-examples/offline-qwen3-asr-decode-files.py \
+    --conv-frontend=./sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25/conv_frontend.onnx \
+    --encoder=./sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25/encoder.int8.onnx \
+    --decoder=./sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25/decoder.int8.onnx \
+    --tokenizer=./sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25/tokenizer \
+    --max-new-tokens=128 \
+    --num-threads=2 \
+    ./sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25/test_wavs/raokouling.wav
 
-rm -rf sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25
+  rm -rf sherpa-onnx-qwen3-asr-0.6B-int8-2026-03-25
+fi
 
 log "test Supertonic TTS"
