+++ .github/scripts/test-rust.sh
@@ -6,8 +6,10 @@ cd rust-api-examples
 
 trap 'bash ../.github/scripts/show-rust-binary-info.sh --all || true' EXIT
 
-./run-qwen3-asr.sh
-rm -rf sherpa-onnx-qwen3-*
+if [[ $SKIP_QWEN3 != "true" ]]; then
+  ./run-qwen3-asr.sh
+  rm -rf sherpa-onnx-qwen3-*
+fi
 
 ./run-audio-tagging-zipformer.sh
 rm -rf sherpa-onnx-zipformer-small-audio-tagging-2024-04-15
